pyproject.toml
[project]
name = "benchmarking-ai-factories"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.1.1",
    "chromadb-client>=1.1.1",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.5",
    "pyyaml>=6.0",
]

[project.scripts]
aif-cli = "src.aif_cli:main"
